backend/main.py
import os
import sys
import torch
import torchvision.transforms as transforms
from PIL import Image
import pickle
import nltk
import string
from nltk.corpus import stopwords
from nltk.stem.porter import PorterStemmer
from fastapi import FastAPI, File, UploadFile, HTTPException
from fastapi.middleware.cors import CORSMiddleware
import io
from pydantic import BaseModel
from contextlib import asynccontextmanager

# ---------------------------
# Setup Paths & NLTK Data
# ---------------------------
# Ensure models directory is in Python path to import model definitions
BASE_DIR = os.path.dirname(os.path.abspath(__file__))
MODELS_DIR = os.path.join(BASE_DIR, "..", "models")
IMAGE_MODEL_DIR = os.path.join(MODELS_DIR, "imageClassification")
SMS_MODEL_DIR = os.path.join(MODELS_DIR, "smsClassifier")

sys.path.insert(0, IMAGE_MODEL_DIR)

try:
    from model import CNNModel
except ImportError:
    pass # Will handle gracefully if missing

# Try to download NLTK data required for preprocessing
try:
    nltk.data.find('corpora/stopwords')
except LookupError:
    nltk.download('stopwords', quiet=True)

try:
    nltk.data.find('tokenizers/punkt')
except LookupError:
    nltk.download('punkt', quiet=True)

# ---------------------------
# Global Models & Device
# ---------------------------
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
cnn_model = None
sms_tfidf = None
sms_model = None

# Preprocessing utilities
transform = transforms.Compose([
    transforms.Resize((64, 64)),
    transforms.ToTensor()
])

ps = PorterStemmer()

def transform_text(text):
    text = text.lower()
    text = nltk.word_tokenize(text)

    y = []
    for i in text:
        if i.isalnum():
            y.append(i)

    text = y[:]
    y.clear()
    
    # ensure stopwords are available
    try:
        stop_words = stopwords.words("english")
    except Exception:
        stop_words = []
        
    for i in text:
        if i not in stop_words and i not in string.punctuation:
            y.append(i)

    text = y[:]
    y.clear()
    for i in text:
        y.append(ps.stem(i))

    return " ".join(y)

# ---------------------------
# Lifespan
# ---------------------------
@asynccontextmanager
async def lifespan(app: FastAPI):
    global cnn_model, sms_tfidf, sms_model
    
    # Load Image Classification Model
    cnn_weights_path = os.path.join(IMAGE_MODEL_DIR, "deep_cnn_model_weights.pth")
    if os.path.exists(cnn_weights_path):
        cnn_model = CNNModel()
        cnn_model.load_state_dict(torch.load(cnn_weights_path, map_location=device))
        cnn_model.to(device)
        cnn_model.eval()
        print("✅ Image Classification model loaded")
    else:
        print(f"⚠️ Warning: Image classification weights not found at {cnn_weights_path}")

    # Load SMS Classification Model
    tfidf_path = os.path.join(SMS_MODEL_DIR, "vectorizer.pkl")
    model_pkl_path = os.path.join(SMS_MODEL_DIR, "model.pkl")
    
    if os.path.exists(tfidf_path) and os.path.exists(model_pkl_path):
        with open(tfidf_path, 'rb') as f:
            sms_tfidf = pickle.load(f)
        with open(model_pkl_path, 'rb') as f:
            sms_model = pickle.load(f)
        print("✅ SMS Classification model loaded")
    else:
        print(f"⚠️ Warning: SMS classification models not found")
        
    yield
    print("Shutting down model API")

# ---------------------------
# FastAPI App setup
# ---------------------------
app = FastAPI(title="DeepFake & SMS classification API", lifespan=lifespan)

# Allow requests from frontend
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"], # For production, restrict to actual frontend URL
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

class TextMessage(BaseModel):
    message: str

# ---------------------------
# Routes
# ---------------------------
@app.get("/")
def read_root():
    return {"status": "ok", "message": "API is running. Use /predict/image or /predict/text."}

@app.post("/predict/image")
async def predict_image(file: UploadFile = File(...)):
    if cnn_model is None:
        raise HTTPException(status_code=503, detail="Image model is not loaded")
    if not file.content_type.startswith("image/"):
        raise HTTPException(status_code=400, detail="Provided file is not an image")
        
    try:
        contents = await file.read()
        image = Image.open(io.BytesIO(contents)).convert("RGB")
        
        # Preprocess
        img_tensor = transform(image).unsqueeze(0).to(device)
        
        # Inference
        with torch.no_grad():
            output = cnn_model(img_tensor)
            probability = output.item()
            
        # Probability >= 0.5 means REAL according to streamlit app
        is_real = probability >= 0.5
        verdict = "Real Media" if is_real else "Deepfake Detected"
        confidence = probability if is_real else (1 - probability)
        
        return {
            "prediction": 0 if is_real else 1, # 0 for real, 1 for deepfake
            "verdict": verdict,
            "confidence": float(confidence),
            "score": int(confidence * 100),
            "metadata": {
                "source": "Deep CNN Analyzer",
                "probability_raw": float(probability)
            }
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/predict/text")
async def predict_text(payload: TextMessage):
    if sms_model is None or sms_tfidf is None:
        raise HTTPException(status_code=503, detail="Text model is not loaded")
        
    if not payload.message or not payload.message.strip():
        raise HTTPException(status_code=400, detail="Empty message provided")
        
    try:
        transformed_sms = transform_text(payload.message)
        vector_input = sms_tfidf.transform([transformed_sms])
        result = sms_model.predict(vector_input)[0]
        
        # Streamlit app indicated result == 1 -> Spam
        prediction = int(result)
        verdict = "Spam/AI-Generated Content Detected" if prediction == 1 else "Safe Content"
        
        # Assuming model.predict_proba is available; if not, we assign default confidence
        try:
            proba = sms_model.predict_proba(vector_input)[0]
            confidence = float(max(proba))
        except AttributeError:
            confidence = 0.95 # Mock high confidence if proba is not supported by standard pickeled model
            
        return {
            "prediction": prediction,
            "verdict": verdict,
            "confidence": confidence,
            "score": int(confidence * 100),
            "metrics": {
                "processed_tokens": len(transformed_sms.split()),
                "suspicious": "High" if prediction == 1 else "Low"
            }
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    import uvicorn
    # run with: uvicorn main:app --reload
    uvicorn.run(app, host="0.0.0.0", port=8000)
