backend/requirements.txt
fastapi[all]>=0.111.0
uvicorn>=0.30.1
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
scikit-learn>=1.0.0
nltk>=3.8.0
